quant_system/requirements.txt
# Core Data & Numerical Computing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Time Series & Financial Data
yfinance==0.2.32
alpha-vantage==2.3.1
pandas-datareader==0.10.0

# Technical Analysis
TA-Lib==0.4.28  # May need manual installation
talib-binary==0.10.5  # Alternative if TA-Lib fails
pandas-ta==0.3.14b  # Pure Python alternative

# Backtesting
backtrader==1.9.94.122
zipline-reloaded==3.1.1

# Portfolio Optimization
cvxpy==1.3.2
plotly==5.17.0

# AI / NLP / LLM
openai==1.3.8
langchain==0.1.0
huggingface-hub==0.19.4
transformers==4.34.1
spacy==3.7.2
nltk==3.8.1

# API & Server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
sqlite3  # Built-in

# Caching & Streaming
redis==5.0.1
kafka-python==2.0.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Logging & Monitoring
python-json-logger==2.0.7
loguru==0.7.2

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==4.0.9
matplotlib==3.8.2
seaborn==0.13.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
